move sample app to examples
