+++ package.json
@@ -24,7 +24,7 @@
 		"format-check": "prettier . --check --ignore-path .gitignore",
 		"format-lint": "pretty-quick --staged --ignore-path .gitignore && lint-staged",
 		"leaks": "bash ./scripts/gitleaks/gitleaks.sh",
-		"lint": "eslint '+(src|test|testUtils)/**/*.+(ts|tsx)' --fix",
+		"lint": "eslint '+(src|test|examples|testUtils)/**/*.+(ts|tsx)' --fix",
 		"prepare": "husky install",
 		"prepublishOnly": "npm run build",
 		"start": "npm run build && rollup -c rollup.config.app.js -w",
